+++ client/src/features/hierarchy/Li.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import RecordLink from '../common/RecordLink'
+
+interface IProps {
+  id: string
+}
+
+const Li: React.FC<IProps> = ({ id }) => (
+  <li>
+    <RecordLink url={id} />
+  </li>
+)
+
+export default Li